+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file. This change
 log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Changed
+- The `=> count` in the `group` operation is now optional. `count` is used by default when omitted:
+```
+email | g: status
+```
 
 ## [0.32.0] - 2026-03-30
 ### Added